+++ util/llbutil/fakedep.go
@@ -1,6 +1,7 @@
 package llbutil
 
 import (
+	"github.com/earthly/earthly/outmon"
 	"github.com/earthly/earthly/util/llbutil/pllb"
 	"github.com/moby/buildkit/client/llb"
 )
@@ -16,6 +17,9 @@ func WithDependency(state pllb.State, depState pllb.State, stateStr, depStr stri
 	// Copy a wildcard that could never exist.
 	// (And allow for the wildcard to match nothing).
 	interm := ScratchWithPlatform()
+	vm := &outmon.VertexMeta{
+		Internal: true,
+	}
 	interm = interm.File(pllb.Copy(
 		depState, "/fake-745cb405-fbfb-4ea7-83b0-a85c26b4aff0-*", "/tmp/",
 		&llb.CopyInfo{
@@ -23,7 +27,11 @@ func WithDependency(state pllb.State, depState pllb.State, stateStr, depStr stri
 			AllowWildcard:       true,
 			AllowEmptyWildcard:  true,
 			CopyDirContentsOnly: true,
-		}), llb.WithCustomNamef("[internal] (fakecopy1) %s depends on %s", stateStr, depStr))
+		}),
+		llb.WithCustomNamef(
+			"%s(fakecopy1) %s depends on %s",
+			vm.ToVertexPrefix(), stateStr, depStr),
+	)
 
 	// Do this again. The extra step is needed to prevent the need for BuildKit
 	// to re-hash the input in certain cases (can be slow if depState is large).
@@ -34,5 +42,9 @@ func WithDependency(state pllb.State, depState pllb.State, stateStr, depStr stri
 			AllowWildcard:       true,
 			AllowEmptyWildcard:  true,
 			CopyDirContentsOnly: true,
-		}), llb.WithCustomNamef("[internal] (fakecopy2) %s depends on %s", stateStr, depStr))
+		}),
+		llb.WithCustomNamef(
+			"%s(fakecopy2) %s depends on %s",
+			vm.ToVertexPrefix(), stateStr, depStr),
+	)
 }